sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentcage"
version = "0.1.0"
description = "Python SDK for the agentcage autonomous security agent platform"
requires-python = ">=3.10"
dependencies = [
    "grpcio>=1.60.0",
    "protobuf>=4.25.0",
]

[project.optional-dependencies]
dev = [
    "grpcio-tools>=1.60.0",
    "pytest>=8.0",
]
